Cloudflare::Rails -> CloudflareRails
