--- lib/cloudflare/rails.rb
@@ -1,10 +1,0 @@
-require "cloudflare/rails/version"
-require 'net/http'
-require 'uri'
-
-module Cloudflare
-  module Rails
-  end
-end
-
-require "cloudflare/rails/railtie"
