--- lib/cloudflare/rails/version.rb
@@ -1,5 +1,0 @@
-module Cloudflare
-  module Rails
-    VERSION = "4.1.0".freeze
-  end
-end
