+++ lib/cloudflare_rails.rb
@@ -1,0 +1,8 @@
+require "cloudflare_rails/version"
+require 'net/http'
+require 'uri'
+
+module CloudflareRails
+end
+
+require "cloudflare_rails/railtie"
