+++ lib/cloudflare_rails/version.rb
@@ -1,0 +1,3 @@
+module CloudflareRails
+  VERSION = "4.1.0".freeze
+end
